refactor: extract action card primitives from AgentWorkspace
